pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "config>=0.5.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "icecream>=2.1.8",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
]
